Added response about validator unavailability
So the problem was in the fact that, all the request were serviced by
chunked response stream. And when the validator became unavailabile,
handler would backoff and retry several times (> a minute in total).
Eventually it would just give up and drop the connection (which nothing
put into into). It's seems to be impossible to return an error in the
middle of succefully started chunked response, as client, after
receiveing some bytes (which describe the error) still awaits for more,
and when the connection is dropped, it considers it to be an error.

So what was done, is to try to perform request to the validator before
initiating a streamed response, and check the response, if
it's not available, return error to the client immediately, otherwise
proceed with initiating chunked response.