Fixed failing tests due to range error
The issue was that helper `account` method generated `AccountData`
values with `len()` ranging from 1, in case if len turned out to be 1,
then `valid_ranges` couldn't generate range, since it also tried to
build range starting from value 1 and ending with whatever is the `len()` of
`AccountData`, 1..1 in case which triggered error.

As a solution, the minimum `AccountData` length was set to 2, since
there's no logic requirements for it to start from 1.

Reworked offsets function in tests to support AccounData of len < 2