Implemented bounded waited queue for rpc requests
It was done via wrapping semaphore type, which was used to control the
limit of concurrent rpc requests, in a new type which also manages the
number of awaiting tasks on that particular semaphore, if that number
becomes more than the preconfigured value, then all the subsequent
requests of that type will be rejected.

Once the task manages to acquire semaphore, it releases its slot in wait
queue, so other tasks will be able to try to acquire wait semaphore.