Implemented optimized filter based caching for gPA (#243)
By introducing the concept of filter supersets and intersections, similar
to ones found in set theory, we are now able to serve requests from
cache for related filters, even if we don't have cache entry for the given filter

Filter superset. One filter is called a superset of the other, if it
selects the data which includes the data selected by other filter.

Ex: Filter(dataSize: X, [memcmp: A]) is the superset of
Filter(dataSize: X, [memcmp: A, memcmp: B]) as it's less restrictive and
will include more data

Filter intersection. An intersection of two filters is another different filter,
which is superset of two previous filters. In different words, it's the
filter which less restrictive than the ones that produced it. Only
filters with 2 memcmp can produce a valid intersection, because otherwise it
wouldn't make sense, and might produce a filter which is superset of all
filters (i.e. empty filter).

Ex: an intersection of Filter(dataSize: X, [memcmp: A, memcmp: B]) and
Filter(dataSize: X, [memcmp: A, memcmp: C]) is Filter(dataSize: X, [memcmp: A]),
which is a superset of filters, that produced it

Cacher will always check for the supersets if no entry is found for the
given filter, and will apply the filter to superset before serving it to
client. If no superset can be found, then it will try to found an
intersection. If found a separate async request will be made to
validator, to fetch accounts for the given intersection filter
(superset), which should prevent future cache misses for similar filters